Extract offline persistence into dedicated module
