+++ assets/main.js
@@ -38,6 +38,14 @@ import {
   bindAudioControls as bindAudioControlElements,
 } from './audioSystem.js';
 import { SimplePlayfield, configurePlayfieldSystem } from './playfield.js';
+import {
+  configureOfflinePersistence,
+  bindOfflineOverlayElements,
+  updatePowderLogDisplay,
+  recordPowderEvent,
+  checkOfflineRewards,
+  markLastActive,
+} from './offlinePersistence.js';
 // Powder tower palette and simulation helpers.
 import {
   DEFAULT_MOTE_PALETTE,
@@ -1520,6 +1528,25 @@ import {
   const powderGlyphColumns = [];
   let powderWallMetrics = null;
 
+  // Wire the standalone offline persistence helpers to the shared gameplay state and utilities.
+  configureOfflinePersistence({
+    formatWholeNumber,
+    formatGameNumber,
+    formatDecimal,
+    formatSignedPercentage,
+    readStorageJson,
+    writeStorageJson,
+    applyPowderGain,
+    notifyIdleTime,
+    getCurrentFluxRate: () => resourceState.fluxRate,
+    onBeforePersist: savePowderCurrency,
+    getCurrentPowderBonuses: () => currentPowderBonuses,
+    powderState,
+    powderConfig,
+    powderElements,
+    updateMoteGemInventoryDisplay,
+  });
+
   function updatePowderGlyphColumns(info = {}) {
     const rows = Number.isFinite(info.rows) && info.rows > 0 ? info.rows : 1;
     const cellSize = Number.isFinite(info.cellSize) && info.cellSize > 0 ? info.cellSize : POWDER_CELL_SIZE_PX;
@@ -1904,11 +1931,7 @@ import {
   let resourceTicker = null;
   let lastResourceTick = 0;
 
-  const powderLog = [];
-  const POWDER_LOG_LIMIT = 6;
-
   const storageKeys = {
-    offline: 'glyph-defense-idle:offline',
     powder: 'glyph-defense-idle:powder',
     audio: AUDIO_SETTINGS_STORAGE_KEY,
     notation: NOTATION_STORAGE_KEY,
@@ -1918,20 +1941,6 @@ import {
 
   let powderCurrency = 0;
   let powderSaveHandle = null;
-
-  const offlineOverlayElements = {
-    container: null,
-    minutes: null,
-    rate: null,
-    total: null,
-    prompt: null,
-  };
-  let offlineOverlayAnimating = false;
-  let offlineOverlayFadeHandle = null;
-  let offlineOverlayPromptHandle = null;
-  let offlineOverlayLastFocus = null;
-  const OFFLINE_OVERLAY_FADE_MS = 220;
-  const OFFLINE_PROMPT_DELAY_MS = 10000;
 
   let powderBasinPulseTimer = null;
 
@@ -5654,929 +5663,6 @@ import {
     }
 
     return amount;
-  }
-
-  function bindOfflineOverlayElements() {
-    offlineOverlayElements.container = document.getElementById('offline-overlay');
-    if (!offlineOverlayElements.container) {
-      return;
-    }
-    offlineOverlayElements.minutes = document.getElementById('offline-minutes');
-    offlineOverlayElements.rate = document.getElementById('offline-rate');
-    offlineOverlayElements.total = document.getElementById('offline-total');
-    offlineOverlayElements.prompt = document.getElementById('offline-prompt');
-
-    offlineOverlayElements.container.addEventListener('pointerdown', (event) => {
-      if (offlineOverlayAnimating) {
-        return;
-      }
-      if (event.pointerType === 'mouse' && event.button !== 0) {
-        return;
-      }
-      hideOfflineOverlay();
-    });
-
-    offlineOverlayElements.container.addEventListener('keydown', (event) => {
-      if (offlineOverlayAnimating) {
-        return;
-      }
-      if (event.key === 'Enter' || event.key === ' ' || event.key === 'Spacebar' || event.key === 'Escape') {
-        event.preventDefault();
-        hideOfflineOverlay();
-      }
-    });
-  }
-
-  function clearOfflineOverlayPrompt() {
-    if (offlineOverlayPromptHandle) {
-      clearTimeout(offlineOverlayPromptHandle);
-      offlineOverlayPromptHandle = null;
-    }
-    if (offlineOverlayElements.prompt) {
-      offlineOverlayElements.prompt.classList.remove('offline-overlay__prompt--visible');
-    }
-  }
-
-  function scheduleOfflineOverlayPrompt() {
-    if (!offlineOverlayElements.prompt) {
-      return;
-    }
-    if (offlineOverlayPromptHandle) {
-      clearTimeout(offlineOverlayPromptHandle);
-      offlineOverlayPromptHandle = null;
-    }
-    offlineOverlayPromptHandle = setTimeout(() => {
-      offlineOverlayPromptHandle = null;
-      if (
-        offlineOverlayElements.prompt &&
-        offlineOverlayElements.container &&
-        offlineOverlayElements.container.classList.contains('active')
-      ) {
-        offlineOverlayElements.prompt.classList.add('offline-overlay__prompt--visible');
-      }
-    }, OFFLINE_PROMPT_DELAY_MS);
-  }
-
-  function animateOfflineNumber(element, target, options = {}) {
-    if (!element) {
-      return Promise.resolve();
-    }
-
-    const settings = {
-      duration: 720,
-      prefix: '',
-      suffix: '',
-      format: formatGameNumber,
-      ...options,
-    };
-
-    const finalValue = Math.max(0, target);
-
-    return new Promise((resolve) => {
-      const start = performance.now();
-
-      const step = (timestamp) => {
-        const elapsed = timestamp - start;
-        const progress = Math.min(1, elapsed / settings.duration);
-        const eased = 1 - (1 - progress) ** 3;
-        const value = progress >= 1 ? finalValue : finalValue * eased;
-        element.textContent = `${settings.prefix}${settings.format(value)}${settings.suffix}`;
-        if (progress < 1) {
-          requestAnimationFrame(step);
-          return;
-        }
-        resolve();
-      };
-
-      requestAnimationFrame(step);
-    });
-  }
-
-  async function showOfflineOverlay(minutes, rate, powder) {
-    const container = offlineOverlayElements.container;
-    if (!container) {
-      return;
-    }
-
-    if (offlineOverlayFadeHandle) {
-      clearTimeout(offlineOverlayFadeHandle);
-      offlineOverlayFadeHandle = null;
-    }
-
-    offlineOverlayAnimating = true;
-    const activeElement = document.activeElement;
-    offlineOverlayLastFocus = activeElement instanceof HTMLElement ? activeElement : null;
-    container.removeAttribute('hidden');
-    container.classList.add('active');
-    container.setAttribute('aria-hidden', 'false');
-
-    clearOfflineOverlayPrompt();
-
-    if (typeof container.focus === 'function') {
-      container.focus({ preventScroll: true });
-    }
-
-    const { minutes: minutesEl, rate: rateEl, total: totalEl } = offlineOverlayElements;
-    if (minutesEl) {
-      minutesEl.textContent = '0';
-    }
-    if (rateEl) {
-      rateEl.textContent = '0';
-    }
-    if (totalEl) {
-      totalEl.textContent = '0';
-    }
-
-    await animateOfflineNumber(minutesEl, minutes, { format: formatWholeNumber });
-    await animateOfflineNumber(rateEl, rate, { format: formatGameNumber });
-    await animateOfflineNumber(totalEl, powder, { format: formatGameNumber });
-
-    offlineOverlayAnimating = false;
-    scheduleOfflineOverlayPrompt();
-  }
-
-  function hideOfflineOverlay() {
-    const container = offlineOverlayElements.container;
-    if (!container) {
-      return;
-    }
-    clearOfflineOverlayPrompt();
-    if (offlineOverlayFadeHandle) {
-      clearTimeout(offlineOverlayFadeHandle);
-      offlineOverlayFadeHandle = null;
-    }
-    offlineOverlayAnimating = true;
-    container.classList.remove('active');
-    container.setAttribute('aria-hidden', 'true');
-    offlineOverlayFadeHandle = setTimeout(() => {
-      container.setAttribute('hidden', '');
-      offlineOverlayFadeHandle = null;
-      offlineOverlayAnimating = false;
-      if (
-        offlineOverlayLastFocus &&
-        typeof offlineOverlayLastFocus.focus === 'function' &&
-        document.contains(offlineOverlayLastFocus)
-      ) {
-        offlineOverlayLastFocus.focus({ preventScroll: true });
-      }
-      offlineOverlayLastFocus = null;
-    }, OFFLINE_OVERLAY_FADE_MS);
-  }
-
-  function checkOfflineRewards() {
-    const savedState = readStorageJson(storageKeys.offline);
-    if (!savedState?.timestamp) {
-      return;
-    }
-
-    const lastActive = Number(savedState.timestamp);
-    if (!Number.isFinite(lastActive)) {
-      return;
-    }
-
-    const now = Date.now();
-    const elapsedMs = Math.max(0, now - lastActive);
-    const minutesAway = Math.floor(elapsedMs / 60000);
-    if (minutesAway <= 0) {
-      return;
-    }
-
-    const storedRate = Number(savedState.powderRate);
-    const effectiveRate = Number.isFinite(storedRate) ? Math.max(0, storedRate) : resourceState.fluxRate;
-    if (effectiveRate <= 0) {
-      return;
-    }
-
-    const powderEarned = minutesAway * effectiveRate;
-    applyPowderGain(powderEarned, { source: 'offline', minutes: minutesAway, rate: effectiveRate });
-    notifyIdleTime(minutesAway * 60000);
-    showOfflineOverlay(minutesAway, effectiveRate, powderEarned);
-  }
-
-  function markLastActive() {
-    savePowderCurrency();
-    writeStorageJson(storageKeys.offline, {
-      timestamp: Date.now(),
-      powderRate: resourceState.fluxRate,
-    });
-  }
-
-  function getCurrentIdleMoteBank() {
-    if (powderSimulation && Number.isFinite(powderSimulation.idleBank)) {
-      return Math.max(0, powderSimulation.idleBank);
-    }
-    if (Number.isFinite(powderState.idleMoteBank)) {
-      return Math.max(0, powderState.idleMoteBank);
-    }
-    return 0;
-  }
-
-  function getCurrentMoteDispenseRate() {
-    if (powderSimulation && Number.isFinite(powderSimulation.idleDrainRate)) {
-      powderState.idleDrainRate = powderSimulation.idleDrainRate;
-      return Math.max(0, powderSimulation.idleDrainRate);
-    }
-    if (Number.isFinite(powderState.idleDrainRate)) {
-      return Math.max(0, powderState.idleDrainRate);
-    }
-    return 0;
-  }
-
-  function formatMoteDispenseRate(rate) {
-    if (!Number.isFinite(rate)) {
-      return '0.00 Motes/sec'; // Present the idle mote flow rate even when values are invalid.
-    }
-    const safeRate = Math.max(0, rate);
-    const formatted = safeRate >= 1
-      ? (Number.isInteger(safeRate) ? formatWholeNumber(safeRate) : formatDecimal(safeRate, 2))
-      : formatDecimal(safeRate, 2);
-    const unit = safeRate === 1 ? 'Mote' : 'Motes';
-    return `${formatted} ${unit}/sec`; // Display the idle mote label consistently across quantities.
-  }
-
-  function updateMoteStatsDisplays() {
-    const storedMotes = getCurrentIdleMoteBank();
-    if (resourceElements.moteStorage) {
-      resourceElements.moteStorage.textContent = `${formatGameNumber(storedMotes)} Motes`; // Reflect the idle mote reserves in the HUD.
-    }
-
-    const dispenseRate = getCurrentMoteDispenseRate();
-    const dispenseLabel = formatMoteDispenseRate(dispenseRate);
-    if (resourceElements.dispenseRate) {
-      resourceElements.dispenseRate.textContent = dispenseLabel;
-    }
-    if (powderElements.dispenseRate) {
-      powderElements.dispenseRate.textContent = dispenseLabel;
-    }
-
-    if (powderElements.idleMultiplier) {
-      const unlocked = getUnlockedAchievementCount();
-      const noun = unlocked === 1 ? 'achievement' : 'achievements';
-      powderElements.idleMultiplier.textContent = `${formatWholeNumber(unlocked)} ${noun}`;
-    }
-  }
-
-  function updateStatusDisplays() {
-    if (resourceElements.score) {
-      const interactive = Boolean(playfield && playfield.isInteractiveLevelActive());
-      const theroValue = interactive ? Math.max(0, Math.round(playfield.energy)) : 0;
-      resourceElements.score.textContent = `${theroValue} ${THERO_SYMBOL}`;
-    }
-    const mintedGlyphs = Math.max(0, Math.floor(getGlyphCurrency()));
-    // Favor the live glyph currency so the HUD mirrors new ℵ progress even if defeat stats lag behind.
-    const glyphsCollected = Math.max(mintedGlyphs, gameStats.enemiesDefeated);
-    const glyphsUnused = Math.max(0, glyphsCollected - gameStats.towersPlaced);
-
-    if (resourceElements.scoreMultiplier) {
-      const multiplier = glyphsCollected && glyphsUnused ? glyphsCollected * glyphsUnused : 0;
-      const display = multiplier ? `×${formatGameNumber(multiplier)}` : '×0';
-      resourceElements.scoreMultiplier.textContent = display;
-    }
-    if (resourceElements.glyphsTotal) {
-      resourceElements.glyphsTotal.textContent = `${formatWholeNumber(glyphsCollected)} Glyphs`;
-    }
-    if (resourceElements.glyphsUnused) {
-      resourceElements.glyphsUnused.textContent = `(${formatWholeNumber(glyphsUnused)} unused)`;
-    }
-    updateMoteStatsDisplays();
-  }
-
-  function updateResourceRates() {
-    currentPowderBonuses = calculatePowderBonuses();
-
-    resourceState.scoreRate = baseResources.scoreRate * currentPowderBonuses.totalMultiplier;
-    const fluxMultiplier = 1 + currentPowderBonuses.sandBonus + currentPowderBonuses.crystalBonus;
-    resourceState.fluxRate = baseResources.fluxRate * fluxMultiplier + getAchievementPowderRate();
-    resourceState.energyRate =
-      baseResources.energyRate * (1 + currentPowderBonuses.duneBonus + currentPowderBonuses.crystalBonus * 0.5);
-
-    updateStatusDisplays();
-  }
-
-  function handleResourceTick(timestamp) {
-    if (!resourceTicker) {
-      return;
-    }
-
-    const activeBeforeUpdate = idleLevelRuns.size;
-    updateIdleRuns(timestamp);
-
-    if (!lastResourceTick) {
-      lastResourceTick = timestamp;
-    }
-
-    const elapsed = Math.max(0, timestamp - lastResourceTick);
-    lastResourceTick = timestamp;
-
-    const effectiveIdleCount = activeBeforeUpdate || idleLevelRuns.size;
-    if (effectiveIdleCount) {
-      notifyIdleTime(elapsed * effectiveIdleCount);
-    }
-
-    const powderGain = resourceState.fluxRate * (elapsed / 60000);
-    if (powderGain > 0) {
-      applyPowderGain(powderGain);
-    }
-
-    if (resourceState.running) {
-      const seconds = elapsed / 1000;
-      resourceState.score += resourceState.scoreRate * seconds;
-    }
-
-    updateStatusDisplays();
-    resourceTicker = requestAnimationFrame(handleResourceTick);
-  }
-
-  function ensureResourceTicker() {
-    if (resourceTicker) {
-      return;
-    }
-    lastResourceTick = 0;
-    resourceTicker = requestAnimationFrame(handleResourceTick);
-  }
-
-  function bindStatusElements() {
-    resourceElements.score = document.getElementById('status-score');
-    resourceElements.scoreMultiplier = document.getElementById('status-score-multiplier');
-    resourceElements.glyphsTotal = document.getElementById('status-glyphs-total');
-    resourceElements.glyphsUnused = document.getElementById('status-glyphs-unused');
-    resourceElements.moteStorage = document.getElementById('status-mote-storage');
-    resourceElements.dispenseRate = document.getElementById('status-dispense-rate');
-    updateStatusDisplays();
-  }
-
-  // Render the mote gem ledger so the motes tab reflects current drop totals.
-  function updateMoteGemInventoryDisplay() {
-    const list = powderElements.gemInventoryList;
-    const emptyState = powderElements.gemInventoryEmpty;
-    if (!list || !emptyState) {
-      return;
-    }
-
-    list.innerHTML = '';
-    const entries = Array.from(moteGemState.inventory.entries());
-    if (!entries.length) {
-      list.hidden = true;
-      list.setAttribute('aria-hidden', 'true');
-      emptyState.hidden = false;
-      emptyState.setAttribute('aria-hidden', 'false');
-      refreshCraftingRecipesDisplay();
-      return;
-    }
-
-    entries.sort((a, b) => {
-      const labelA = a[1]?.label || '';
-      const labelB = b[1]?.label || '';
-      return labelA.localeCompare(labelB);
-    });
-
-    const fragment = document.createDocumentFragment();
-    entries.forEach(([typeKey, record]) => {
-      if (!record) {
-        return;
-      }
-
-      const item = document.createElement('li');
-      item.className = 'powder-gem-inventory__item';
-
-      const labelGroup = document.createElement('span');
-      labelGroup.className = 'powder-gem-inventory__label';
-
-      const swatch = document.createElement('span');
-      swatch.className = 'powder-gem-inventory__swatch';
-      const color = getMoteGemColor(typeKey);
-      if (color) {
-        if (Number.isFinite(color.hue)) {
-          swatch.style.setProperty('--gem-hue', `${Math.round(color.hue)}`);
-        }
-        if (Number.isFinite(color.saturation)) {
-          swatch.style.setProperty('--gem-saturation', `${Math.round(color.saturation)}%`);
-        }
-        if (Number.isFinite(color.lightness)) {
-          swatch.style.setProperty('--gem-lightness', `${Math.round(color.lightness)}%`);
-        }
-      }
-
-      const name = document.createElement('span');
-      name.className = 'powder-gem-inventory__name';
-      name.textContent = record.label || 'Mote Gem';
-
-      labelGroup.append(swatch, name);
-
-      const count = document.createElement('span');
-      count.className = 'powder-gem-inventory__count';
-      count.textContent = formatGameNumber(Math.max(0, record.total || 0));
-
-      item.append(labelGroup, count);
-      fragment.append(item);
-    });
-
-    list.hidden = false;
-    list.setAttribute('aria-hidden', 'false');
-    emptyState.hidden = true;
-    emptyState.setAttribute('aria-hidden', 'true');
-    list.append(fragment);
-    refreshCraftingRecipesDisplay();
-  }
-
-  function bindPowderControls() {
-    if (!powderState.fluidUnlocked && gameStats.powderSigilsReached >= (powderConfig.fluidUnlockSigils || Infinity)) {
-      powderState.fluidUnlocked = true;
-    }
-    powderElements.sandfallFormula = document.getElementById('powder-sandfall-formula');
-    powderElements.sandfallNote = document.getElementById('powder-sandfall-note');
-    powderElements.sandfallButton = document.querySelector('[data-powder-action="sandfall"]');
-
-    powderElements.duneFormula = document.getElementById('powder-dune-formula');
-    powderElements.duneNote = document.getElementById('powder-dune-note');
-    powderElements.duneButton = document.querySelector('[data-powder-action="dune"]');
-
-    powderElements.crystalFormula = document.getElementById('powder-crystal-formula');
-    powderElements.crystalNote = document.getElementById('powder-crystal-note');
-    powderElements.crystalButton = document.querySelector('[data-powder-action="crystal"]');
-
-    powderElements.simulationCanvas = document.getElementById('powder-canvas');
-    powderElements.basin = document.getElementById('powder-basin');
-    powderElements.viewport = document.getElementById('powder-viewport');
-    powderElements.leftWall = document.getElementById('powder-wall-left');
-    powderElements.rightWall = document.getElementById('powder-wall-right');
-    powderElements.leftHitbox = document.getElementById('powder-wall-hitbox-left');
-    powderElements.rightHitbox = document.getElementById('powder-wall-hitbox-right');
-    powderElements.wallMarker = document.getElementById('powder-wall-marker');
-    powderElements.crestMarker = document.getElementById('powder-crest-marker');
-    powderElements.wallGlyphColumns = Array.from(
-      document.querySelectorAll('[data-powder-glyph-column]'),
-    );
-    powderGlyphColumns.length = 0;
-    powderElements.wallGlyphColumns.forEach((columnEl) => {
-      if (!columnEl) {
-        return;
-      }
-      columnEl.innerHTML = '';
-      powderGlyphColumns.push({ element: columnEl, glyphs: new Map() });
-    });
-    applyPowderViewportTransform(powderState.viewTransform);
-
-    function initializePowderViewInteraction() {
-      const canvas = powderElements.simulationCanvas;
-      if (!canvas) {
-        return;
-      }
-      if (!powderState.viewInteraction) {
-        powderState.viewInteraction = {
-          pointerId: null,
-          startX: 0,
-          startY: 0,
-          startCenter: { x: 0.5, y: 0.5 },
-          dragging: false,
-          suppressClick: false,
-          activePointers: new Map(),
-          pinchState: null,
-          isPinchZooming: false,
-          handlers: null,
-        };
-      }
-      const viewInteraction = powderState.viewInteraction;
-      viewInteraction.activePointers = new Map();
-      viewInteraction.pinchState = null;
-      viewInteraction.isPinchZooming = false;
-      if (!viewInteraction.handlers) {
-        const performPinchZoom = () => {
-          const pointers = Array.from(viewInteraction.activePointers.values());
-          if (pointers.length < 2 || !powderSimulation) {
-            viewInteraction.pinchState = null;
-            viewInteraction.isPinchZooming = false;
-            return;
-          }
-          const [first, second] = pointers;
-          const dx = first.clientX - second.clientX;
-          const dy = first.clientY - second.clientY;
-          const distance = Math.hypot(dx, dy);
-          if (!Number.isFinite(distance) || distance <= 0) {
-            return;
-          }
-          const midpoint = {
-            clientX: (first.clientX + second.clientX) / 2,
-            clientY: (first.clientY + second.clientY) / 2,
-          };
-          if (!viewInteraction.pinchState || !Number.isFinite(viewInteraction.pinchState.startDistance) || viewInteraction.pinchState.startDistance <= 0) {
-            viewInteraction.pinchState = {
-              startDistance: distance,
-              startScale: powderSimulation.viewScale || 1,
-            };
-            return;
-          }
-          const baseDistance = viewInteraction.pinchState.startDistance;
-          const baseScale = viewInteraction.pinchState.startScale || powderSimulation.viewScale || 1;
-          if (!Number.isFinite(baseDistance) || baseDistance <= 0) {
-            return;
-          }
-          const targetScale = (distance / baseDistance) * baseScale;
-          powderSimulation.setZoom(targetScale, midpoint);
-          viewInteraction.pinchState.startScale = powderSimulation.viewScale;
-          viewInteraction.pinchState.startDistance = distance;
-          viewInteraction.isPinchZooming = true;
-          viewInteraction.suppressClick = true;
-        };
-
-        viewInteraction.handlers = {
-          pointerDown: (event) => {
-            if (!powderSimulation) {
-              return;
-            }
-            if (event.pointerType === 'touch') {
-              viewInteraction.activePointers.set(event.pointerId, {
-                clientX: event.clientX,
-                clientY: event.clientY,
-              });
-              if (viewInteraction.activePointers.size >= 2) {
-                viewInteraction.pointerId = null;
-                viewInteraction.dragging = false;
-                viewInteraction.pinchState = null;
-                return;
-              }
-            }
-            if (event.pointerType !== 'touch' && event.button !== 0) {
-              return;
-            }
-            viewInteraction.pointerId = event.pointerId;
-            viewInteraction.startX = event.clientX;
-            viewInteraction.startY = event.clientY;
-            viewInteraction.startCenter = {
-              ...(powderSimulation.viewCenterNormalized || { x: 0.5, y: 0.5 }),
-            };
-            viewInteraction.dragging = false;
-            viewInteraction.suppressClick = false;
-            viewInteraction.pinchState = null;
-            viewInteraction.isPinchZooming = false;
-            if (typeof canvas.setPointerCapture === 'function') {
-              try {
-                canvas.setPointerCapture(event.pointerId);
-              } catch (error) {
-                // Ignore capture errors so gestures still proceed without it.
-              }
-            }
-          },
-          pointerMove: (event) => {
-            if (!powderSimulation) {
-              return;
-            }
-            if (event.pointerType === 'touch') {
-              viewInteraction.activePointers.set(event.pointerId, {
-                clientX: event.clientX,
-                clientY: event.clientY,
-              });
-              if (viewInteraction.activePointers.size >= 2) {
-                if (typeof event.preventDefault === 'function') {
-                  event.preventDefault();
-                }
-                performPinchZoom();
-                return;
-              }
-            }
-            if (viewInteraction.pointerId !== event.pointerId) {
-              return;
-            }
-            const dx = event.clientX - viewInteraction.startX;
-            const dy = event.clientY - viewInteraction.startY;
-            const distance = Math.hypot(dx, dy);
-            if (!viewInteraction.dragging && distance > 6) {
-              viewInteraction.dragging = true;
-              viewInteraction.suppressClick = true;
-            }
-            if (!viewInteraction.dragging) {
-              return;
-            }
-            if (typeof event.preventDefault === 'function') {
-              event.preventDefault();
-            }
-            const width = powderSimulation.width || powderSimulation.canvas?.clientWidth || 0;
-            const height = powderSimulation.height || powderSimulation.canvas?.clientHeight || 0;
-            const scale = powderSimulation.viewScale || 1;
-            if (!width || !height) {
-              return;
-            }
-            const nextCenter = {
-              x: viewInteraction.startCenter.x - dx / (scale * width),
-              y: viewInteraction.startCenter.y - dy / (scale * height),
-            };
-            powderSimulation.setViewCenterNormalized(nextCenter);
-          },
-          pointerUp: (event) => {
-            if (event.pointerType === 'touch') {
-              viewInteraction.activePointers.delete(event.pointerId);
-              if (viewInteraction.activePointers.size < 2) {
-                viewInteraction.pinchState = null;
-                viewInteraction.isPinchZooming = false;
-              }
-            }
-            if (viewInteraction.pointerId === event.pointerId) {
-              if (viewInteraction.dragging || viewInteraction.isPinchZooming) {
-                viewInteraction.suppressClick = true;
-              }
-              viewInteraction.pointerId = null;
-              viewInteraction.dragging = false;
-            }
-            if (typeof canvas.releasePointerCapture === 'function') {
-              try {
-                canvas.releasePointerCapture(event.pointerId);
-              } catch (error) {
-                // Ignore browsers that throw if the pointer was not captured.
-              }
-            }
-          },
-          wheel: (event) => {
-            if (!powderSimulation) {
-              return;
-            }
-            if (typeof event.preventDefault === 'function') {
-              event.preventDefault();
-            }
-            const delta = Number.isFinite(event.deltaY) ? event.deltaY : 0;
-            const factor = Math.exp(-delta * 0.0015);
-            powderSimulation.applyZoomFactor(factor, event);
-            viewInteraction.suppressClick = true;
-          },
-          click: (event) => {
-            if (!viewInteraction.suppressClick) {
-              return;
-            }
-            viewInteraction.suppressClick = false;
-            event.preventDefault();
-            event.stopPropagation();
-          },
-        };
-      }
-
-      if (!canvas.dataset.powderViewBound) {
-        canvas.addEventListener('pointerdown', viewInteraction.handlers.pointerDown);
-        canvas.addEventListener('pointermove', viewInteraction.handlers.pointerMove);
-        canvas.addEventListener('pointerup', viewInteraction.handlers.pointerUp);
-        canvas.addEventListener('pointercancel', viewInteraction.handlers.pointerUp);
-        canvas.addEventListener('lostpointercapture', viewInteraction.handlers.pointerUp);
-        canvas.addEventListener('wheel', viewInteraction.handlers.wheel, { passive: false });
-        canvas.addEventListener('click', viewInteraction.handlers.click, true);
-        canvas.dataset.powderViewBound = 'true';
-      }
-    }
-
-    updatePowderGlyphColumns({
-      rows: 1,
-      cellSize: POWDER_CELL_SIZE_PX,
-      scrollOffset: 0,
-      highestNormalized: 0,
-      totalNormalized: 0,
-    });
-
-    powderElements.totalMultiplier = document.getElementById('powder-total-multiplier');
-    powderElements.sandBonusValue = document.getElementById('powder-sand-bonus');
-    powderElements.duneBonusValue = document.getElementById('powder-dune-bonus');
-    powderElements.crystalBonusValue = document.getElementById('powder-crystal-bonus');
-    powderElements.stockpile = document.getElementById('powder-stockpile');
-    powderElements.idleMultiplier = document.getElementById('powder-idle-multiplier');
-    powderElements.dispenseRate = document.getElementById('powder-dispense-rate');
-
-    // Bind the gem inventory UI so the motes tab mirrors collected mote gem totals.
-    powderElements.gemInventoryList = document.getElementById('powder-gem-inventory');
-    powderElements.gemInventoryEmpty = document.getElementById('powder-gem-empty');
-    powderElements.craftingButton = document.getElementById('open-crafting-menu');
-
-    if (powderElements.craftingButton) {
-      powderElements.craftingButton.addEventListener('click', () => {
-        openCraftingOverlay();
-      });
-    }
-
-    updateMoteGemInventoryDisplay();
-
-    powderElements.ledgerBaseScore = document.getElementById('powder-ledger-base-score');
-    powderElements.ledgerCurrentScore = document.getElementById('powder-ledger-current-score');
-    powderElements.ledgerFlux = document.getElementById('powder-ledger-flux');
-    powderElements.ledgerEnergy = document.getElementById('powder-ledger-energy');
-
-    powderElements.sigilEntries = Array.from(
-      document.querySelectorAll('[data-sigil-threshold]'),
-    );
-
-    powderElements.logList = document.getElementById('powder-log');
-    powderElements.logEmpty = document.getElementById('powder-log-empty');
-    powderElements.modeToggle = document.getElementById('powder-mode-toggle');
-
-    if (powderElements.sandfallButton) {
-      powderElements.sandfallButton.addEventListener('click', toggleSandfallStability);
-    }
-
-    if (powderElements.duneButton) {
-      powderElements.duneButton.addEventListener('click', surveyRidgeHeight);
-    }
-
-    if (powderElements.crystalButton) {
-      powderElements.crystalButton.addEventListener('click', chargeCrystalMatrix);
-    }
-
-    if (powderElements.modeToggle) {
-      powderElements.modeToggle.addEventListener('click', () => {
-        handlePowderModeToggle();
-      });
-      updatePowderModeButton();
-    }
-
-    if (powderElements.simulationCanvas && !powderSimulation) {
-      const { left: leftInset, right: rightInset } = getPowderWallInsets();
-      powderSimulation = new PowderSimulation({
-        canvas: powderElements.simulationCanvas,
-        cellSize: POWDER_CELL_SIZE_PX,
-        grainSizes: [1, 2, 3],
-        scrollThreshold: 0.75,
-        wallInsetLeft: leftInset,
-        wallInsetRight: rightInset,
-        wallGapCells: powderConfig.wallBaseGapMotes,
-        maxDuneGain: powderConfig.simulatedDuneGainMax,
-        idleDrainRate: powderState.idleDrainRate,
-        motePalette: powderState.motePalette,
-        onHeightChange: handlePowderHeightChange,
-        onWallMetricsChange: handlePowderWallMetricsChange,
-        onViewTransformChange: handlePowderViewTransformChange,
-      });
-      sandSimulation = powderSimulation;
-      powderState.idleDrainRate = powderSimulation.idleDrainRate;
-      powderSimulation.setFlowOffset(powderState.sandOffset);
-      powderSimulation.start();
-      initializePowderViewInteraction();
-      if (powderBasinObserver && typeof powderBasinObserver.disconnect === 'function') {
-        powderBasinObserver.disconnect();
-        powderBasinObserver = null;
-      }
-      if (typeof ResizeObserver === 'function' && powderElements.basin) {
-        powderBasinObserver = new ResizeObserver(() => {
-          if (powderSimulation) {
-            powderSimulation.handleResize();
-          }
-        });
-        powderBasinObserver.observe(powderElements.basin);
-      }
-      handlePowderHeightChange(powderSimulation.getStatus());
-      updatePowderWallGapFromGlyphs(powderState.wallGlyphsLit || 0);
-      initializePowderViewInteraction();
-      handlePowderViewTransformChange(powderSimulation.getViewTransform());
-      syncPowderWallVisuals();
-      updatePowderHitboxVisibility();
-      flushPendingMoteDrops();
-    }
-
-    flushPendingMoteDrops();
-
-    updatePowderStockpileDisplay();
-    updateMoteStatsDisplays();
-    updatePowderModeButton();
-  }
-
-  function updatePowderLedger() {
-    if (powderElements.ledgerBaseScore) {
-      powderElements.ledgerBaseScore.textContent = `${formatGameNumber(
-        baseResources.scoreRate,
-      )} Σ/s`;
-    }
-
-    if (powderElements.ledgerCurrentScore) {
-      powderElements.ledgerCurrentScore.textContent = `${formatGameNumber(
-        resourceState.scoreRate,
-      )} Σ/s`;
-    }
-
-    if (powderElements.ledgerFlux) {
-      powderElements.ledgerFlux.textContent = `+${formatGameNumber(
-        resourceState.fluxRate,
-      )} Mote Gems/min`;
-    }
-
-    if (powderElements.ledgerEnergy) {
-      powderElements.ledgerEnergy.textContent = `+${formatGameNumber(
-        resourceState.energyRate,
-      )} TD/s`;
-    }
-  }
-
-  function updatePowderLogDisplay() {
-    if (!powderElements.logList || !powderElements.logEmpty) {
-      return;
-    }
-
-    powderElements.logList.innerHTML = '';
-
-    if (!powderLog.length) {
-      powderElements.logList.setAttribute('hidden', '');
-      powderElements.logEmpty.hidden = false;
-      return;
-    }
-
-    powderElements.logList.removeAttribute('hidden');
-    powderElements.logEmpty.hidden = true;
-
-    const fragment = document.createDocumentFragment();
-    powderLog.forEach((entry) => {
-      const item = document.createElement('li');
-      item.textContent = entry;
-      fragment.append(item);
-    });
-    powderElements.logList.append(fragment);
-  }
-
-  function recordPowderEvent(type, context = {}) {
-    let entry = '';
-
-    switch (type) {
-      case 'sand-stabilized': {
-        entry = `Sandfall stabilized · Mote bonus ${formatSignedPercentage(
-          currentPowderBonuses.sandBonus,
-        )}.`;
-        break;
-      }
-      case 'sand-released': {
-        entry = 'Sandfall released · Flow returns to natural mote drift.';
-        break;
-      }
-      case 'dune-raise': {
-        const { height = powderState.duneHeight } = context;
-        const logValue = Math.log2(height + 1);
-        entry = `Dune surveyed · h = ${height}, Δm = ${formatDecimal(logValue, 2)}.`;
-        break;
-      }
-      case 'dune-max': {
-        entry = 'Dune survey halted · Ridge already at maximum elevation.';
-        break;
-      }
-      case 'crystal-charge': {
-        const { charges = powderState.charges } = context;
-        entry = `Crystal lattice charged (${charges}/3) · Resonance rising.`;
-        break;
-      }
-      case 'crystal-release': {
-        const { pulseBonus = 0 } = context;
-        entry = `Crystal pulse released · Σ surged ${formatSignedPercentage(pulseBonus)}.`;
-        break;
-      }
-      case 'achievement-unlocked': {
-        const { title = 'Achievement' } = context;
-        entry = `${title} seal unlocked · +1 Motes/min secured.`;
-        break;
-      }
-      case 'offline-reward': {
-        const { minutes = 0, rate = 0, powder = 0 } = context;
-        const minutesLabel = formatWholeNumber(minutes);
-        entry = `Idle harvest · ${minutesLabel}m × ${formatGameNumber(rate)} = +${formatGameNumber(
-          powder,
-        )} Motes.`;
-        break;
-      }
-      case 'developer-adjust': {
-        const { field = 'value', value = 0 } = context;
-        const fieldLabels = {
-          'idle-mote-bank': 'Idle mote bank',
-          'idle-mote-rate': 'Idle mote fall rate',
-          'base-start-thero': 'Base start þ',
-          glyphs: 'Glyph reserves',
-        };
-        const label = fieldLabels[field] || field;
-        entry = `Developer adjusted ${label} → ${formatGameNumber(Number(value) || 0)}.`;
-        break;
-      }
-      case 'mote-gem-collected': {
-        const { type = 'Mote Gem', value = 0 } = context;
-        // Record the mote gem pickup while refreshing the visible inventory ledger.
-        updateMoteGemInventoryDisplay();
-        entry = `${type} cluster secured · +${formatGameNumber(Math.max(0, value || 0))}.`;
-        break;
-      }
-      case 'mode-switch': {
-        const { mode = powderState.simulationMode, label } = context;
-        const normalizedMode = mode === 'fluid' ? 'fluid' : 'sand';
-        const modeLabel =
-          normalizedMode === 'fluid'
-            ? label || powderState.fluidProfileLabel || 'Fluid Study'
-            : 'Powderfall Study';
-        entry = `Simulation mode changed · ${modeLabel} engaged.`;
-        break;
-      }
-      case 'fluid-unlocked': {
-        const { threshold = powderConfig.fluidUnlockSigils || 0 } = context;
-        const unitLabel = threshold === 1 ? 'Glyph' : 'Glyphs';
-        entry = `Fluid resonance unlocked · Requires ${threshold} ${unitLabel}.`;
-        break;
-      }
-      default:
-        break;
-    }
-
-    if (!entry) {
-      return;
-    }
-
-    powderLog.unshift(entry);
-    if (powderLog.length > POWDER_LOG_LIMIT) {
-      powderLog.length = POWDER_LOG_LIMIT;
-    }
-    updatePowderLogDisplay();
   }
 
   // Connect mote gem logging to the powder handlers once both helpers are defined.
